requirements.txt
# Email Agent Requirements
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
pydantic>=2.5.0
regex>=2023.10.3
spacy>=3.7.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
python-dateutil>=2.8.2
pytz>=2023.3
slack-sdk>=3.23.0
notion-client>=2.2.0
pytesseract>=0.3.10
Pillow>=10.1.0
PyPDF2>=3.0.1
